Term3/Project-3-System_Integration/tieinfighter-final-master/requirements-2.txt
actionlib==1.11.13
angles==1.9.11
attrdict==2.0.0
backports.weakref==1.0.post1
bleach==1.5.0
bondpy==1.7.20
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
catkin==0.6.19
click==6.7
cv-bridge==1.11.16
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
dynamic-reconfigure==1.5.49
eventlet==0.19.0
Flask==0.11.1
funcsigs==1.0.2
gazebo-plugins==2.4.16
gazebo-ros==2.4.16
gencpp==0.5.5
genlisp==0.4.15
genmsg==0.5.8
genpy==0.5.10
greenlet==0.4.14
h5py==2.6.0
html5lib==0.9999999
image-geometry==1.11.16
interactive-markers==1.11.3
itsdangerous==0.24
Jinja2==2.10
Keras==2.0.8
laser-geometry==1.6.4
Markdown==2.6.11
MarkupSafe==1.0
message-filters==1.11.21
mock==2.0.0
numpy==1.13.1
p4python==2017.1.1526044
pbr==4.1.1
Pillow==5.2.0
pluginlib==1.10.7
protobuf==3.6.0
python-engineio==2.2.0
python-qt-binding==0.2.19
python-socketio==1.6.1
PyYAML==3.13
qt-dotgraph==0.2.33
qt-gui==0.2.33
qt-gui-cpp==0.2.33
qt-gui-py-common==0.2.33
resource-retriever==1.11.8
rosbag==1.11.21
rosboost-cfg==1.11.14
rosclean==1.11.14
roscreate==1.11.14
rosgraph==1.11.21
roslaunch==1.11.21
roslib==1.11.14
roslint==0.10.0
roslz4==1.11.21
rosmake==1.11.14
rosmaster==1.11.21
rosmsg==1.11.21
rosnode==1.11.21
rosparam==1.11.21
rospy==1.11.21
rosservice==1.11.21
rostest==1.11.21
rostopic==1.11.21
rosunit==1.11.14
roswtf==1.11.21
rqt-action==0.4.9
rqt-bag==0.4.8
rqt-bag-plugins==0.4.8
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.8
rqt-gui==0.4.8
rqt-gui-py==0.4.8
rqt-image-view==0.4.9
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.4.8
rqt-py-console==0.4.8
rqt-reconfigure==0.4.10
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.7
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.5.8
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rviz==1.11.19
scipy==0.19.1
sensor-msgs==1.11.10
six==1.11.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.7.20
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.17
tf2-kdl==0.5.17
tf2-py==0.5.17
tf2-ros==0.5.17
tf2-sensor-msgs==0.5.17
Werkzeug==0.14.1
xacro==1.9.5
